fix: implemented the required changes
