+++ src/helpers/Errors.interfaces.ts
@@ -1,0 +1,10 @@
+export interface HttpExceptionResponse {
+  statusCode: number;
+  error: string;
+}
+
+export interface CustomHttpExceptionResponse extends HttpExceptionResponse {
+  path: string;
+  method: string;
+  timeStamp: Date;
+}
